Coding 11

+++ controllers/zip_del_data.php
@@ -3,33 +3,46 @@
 include("../helpers/mysqli_connect.php");
 
 if(!empty($_POST['id'])){
-    $data;
     try {
-        //del data from the database
-        $sql = "DELETE FROM contents WHERE content_id = {$_POST['id']} AND content_extension = 'zip' ";
-        $conn->query($sql);
-
-        //re order data from the database
-        $sql = "SELECT content_id FROM contents WHERE content_extension = 'zip' ORDER BY content_order";
+        $sql = "SELECT CONCAT(content_url, content_name, '.',content_extension) as contentPath FROM contents WHERE content_id = {$_POST['id']} AND content_extension = 'zip' ";
         $result = $conn->query($sql);
         if ($result->num_rows > 0) {
             //output data of each row
-            $rowNum = 0;
             while($row = $result->fetch_assoc()) {
-                $rowNum = $rowNum + 1;
-            
-                $sql = "UPDATE contents SET content_order = {$rowNum} WHERE content_id = {$row['content_id']} AND content_extension = 'zip' ";
-                $conn->query($sql);
+                $content_path = $row['contentPath'];
             }
+        }else {
+            echo json_encode(false);
+            return;
         }
 
-        $data = true;
+        $flgDelete = unlink($content_path);
+        if($flgDelete)
+        {
+            //del data from the database
+            $sql = "DELETE FROM contents WHERE content_id = {$_POST['id']} AND content_extension = 'zip' ";
+            $conn->query($sql);
+
+            //re order data from the database
+            $sql = "SELECT content_id FROM contents WHERE content_extension = 'zip' ORDER BY content_order";
+            $result = $conn->query($sql);
+            if ($result->num_rows > 0) {
+                //output data of each row
+                $rowNum = 0;
+                while($row = $result->fetch_assoc()) {
+                    $rowNum = $rowNum + 1;
+                
+                    $sql = "UPDATE contents SET content_order = {$rowNum} WHERE content_id = {$row['content_id']} AND content_extension = 'zip' ";
+                    $conn->query($sql);
+                }
+            }
+
+            echo json_encode(true);
+        }
 
     } catch (Exception $e) {
-        $data = false;
+        echo json_encode(false);
     }
 }
 
-//returns data as JSON format
-echo json_encode($data);
 ?>
